# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: BSY MIS Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /applications:
    get:
      operationId: listApplications
      tags: [applications]
      summary: List applications (filtered by role)
      parameters:
        - name: status
          in: query
          schema:
            type: ["string", "null"]
        - name: page
          in: query
          schema:
            type: ["integer", "null"]
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of applications
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicationListResponse"
    post:
      operationId: createApplication
      tags: [applications]
      summary: Create new application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
  /applications/{id}:
    get:
      operationId: getApplication
      tags: [applications]
      summary: Get application by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
    patch:
      operationId: updateApplication
      tags: [applications]
      summary: Update application
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
  /applications/{id}/submit:
    post:
      operationId: submitApplication
      tags: [applications]
      summary: Submit application for review
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
  /applications/{id}/documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: Get documents for application
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
    post:
      operationId: uploadDocument
      tags: [documents]
      summary: Upload a document for application
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentInput"
      responses:
        "201":
          description: Document uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
  /applications/{id}/review:
    post:
      operationId: reviewApplication
      tags: [reviews]
      summary: GIMABA review action on application
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "200":
          description: Review recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
  /applications/{id}/sir:
    get:
      operationId: getSir
      tags: [sir]
      summary: Get SIR report for application
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: SIR report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SirReport"
    post:
      operationId: submitSir
      tags: [sir]
      summary: Protection Officer submits SIR report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SirInput"
      responses:
        "201":
          description: SIR submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SirReport"
...
components:
  schemas:
...
    Application:
      type: object
      properties:
        id:
          type: integer
        applicationNumber:
          type: string
        applicantId:
          type: integer
        childName:
          type: string
        childDob:
          type: string
        childGender:
          type: string
        aadhaarNumber:
          type: string
        casteCategory:
          type: string
        beneficiaryCategory:
          type: string
        guardianName:
          type: string
        guardianMobile:
          type: string
        annualIncome:
          type: number
        district:
          type: string
        taluka:
          type: string
        address:
          type: string
        sansthaId:
          type: ["integer", "null"]
        facilitatorType:
          type: ["string", "null"]
        financialYear:
          type: string
        status:
          type: string
          enum: [draft, submitted, under_review, pending_documents, forwarded_to_po, sir_pending, pwc_scheduled, approved, rejected]
        aadhaarSeeded:
          type: boolean
        submittedAt:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, applicationNumber, applicantId, childName, childDob, childGender, aadhaarNumber, casteCategory, beneficiaryCategory, guardianName, guardianMobile, annualIncome, district, taluka, address, financialYear, status, aadhaarSeeded, createdAt, updatedAt]
    ApplicationListResponse:
      type: object
      properties:
        applications:
          type: array
          items:
            $ref: "#/components/schemas/Application"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [applications, total, page, limit]
    ApplicationInput:
      type: object
      properties:
        childName:
          type: string
        childDob:
          type: string
        childGender:
          type: string
        aadhaarNumber:
          type: string
        casteCategory:
          type: string
        beneficiaryCategory:
          type: string
        guardianName:
          type: string
        guardianMobile:
          type: string
        annualIncome:
          type: number
        district:
          type: string
        taluka:
          type: string
        address:
          type: string
        sansthaId:
          type: ["integer", "null"]
        facilitatorType:
          type: ["string", "null"]
        bankAccountNumber:
          type: string
        ifscCode:
          type: string
        accountHolderName:
          type: string
        fatherStatus:
          type: string
        motherStatus:
          type: string
        schoolName:
          type: ["string", "null"]
        currentClass:
          type: ["string", "null"]
        pincode:
          type: string
      required: [childName, childDob, childGender, aadhaarNumber, casteCategory, beneficiaryCategory, guardianName, guardianMobile, annualIncome, district, taluka, address, bankAccountNumber, ifscCode, accountHolderName, fatherStatus, motherStatus, pincode]
    ApplicationUpdate:
      type: object
      properties:
        childName:
          type: string
        guardianName:
          type: string
        address:
          type: string
        annualIncome:
          type: number
        bankAccountNumber:
          type: string
        ifscCode:
          type: string
        accountHolderName:
          type: string
        schoolName:
          type: ["string", "null"]
        currentClass:
          type: ["string", "null"]
    Document:
      type: object
      properties:
        id:
          type: integer
        applicationId:
          type: integer
        docType:
          type: string
        fileUrl:
          type: string
        fileName:
          type: string
        uploadedAt:
          type: string
        verifiedBy:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, verified, rejected]
      required: [id, applicationId, docType, fileUrl, fileName, uploadedAt, status]
    DocumentInput:
      type: object
      properties:
        docType:
          type: string
        fileUrl:
          type: string
        fileName:
          type: string
      required: [docType, fileUrl, fileName]
    Review:
      type: object
      properties:
        id:
          type: integer
        applicationId:
          type: integer
        reviewerId:
          type: integer
        action:
          type: string
          enum: [approved, rejected, forwarded_to_po, requested_documents]
        reason:
          type: ["string", "null"]
        actionDate:
          type: string
      required: [id, applicationId, reviewerId, action, actionDate]
    ReviewInput:
      type: object
      properties:
        action:
          type: string
          enum: [approved, rejected, forwarded_to_po, requested_documents]
        reason:
          type: ["string", "null"]
      required: [action]
    SirReport:
      type: object
      properties:
        id:
          type: integer
        applicationId:
          type: integer
        poId:
          type: integer
        visitDate:
          type: string
        geoLat:
          type: ["number", "null"]
        geoLng:
          type: ["number", "null"]
        photoUrls:
          type: array
          items:
            type: string
        recommendation:
          type: string
          enum: [recommended, not_recommended, further_inquiry]
        observations:
          type: string
        submittedAt:
          type: string
      required: [id, applicationId, poId, visitDate, recommendation, observations, submittedAt, photoUrls]
    SirInput:
      type: object
      properties:
        visitDate:
          type: string
        geoLat:
          type: ["number", "null"]
        geoLng:
          type: ["number", "null"]
        photoUrls:
          type: array
          items:
            type: string
        recommendation:
          type: string
          enum: [recommended, not_recommended, further_inquiry]
        observations:
          type: string
      required: [visitDate, recommendation, observations, photoUrls]
...