# lib/api-zod/src/generated/types/applicationListResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BSY MIS Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { Application } from './application';

export interface ApplicationListResponse {
  applications: Application[];
  total: number;
  page: number;
  limit: number;
}